requirements.txt
easygui==0.98.3
Pillow==10.0.0
pygame==2.5.1
python-dotenv==1.0.0

pip~=23.1.2
requests~=2.31.0
yarg~=0.1.9
mysqlclient~=2.2.0
setuptools~=65.5.0